Create environment.yml for conda env
As `pytorch` is much easier to install when using `conda`, we abandon to
use `pip` to install python packages. Hence, the file `requirements.txt`
is replaced with a file `environment.yml` which is used for `conda`.

The example command is as follows:
```bash
conda env create --name fym --file environment.yml
```