Add a rotation module
This module replaces `scipy`'s `rotation` module which is revealed to
be very slow. In simple experiments, it turns out that our module is
twice as faster as `scipy`'s. This is related to #84

This module is a set of functions which are python versions of MATLAB's
`angle2dcm` and `dcm2angle`. Currently, this module only contains a
rotation type(`rtype`) `zyx`, but other types can be easily added.