Add eager_stop and logger_callback in BaseEnv
`eager_stop` is a function called in `update` method.
It takes two arguments: `t_hist` and `ode_hist`,
and returns a `True` flag that the simulation should finished before
the clock hits the final time.

`logger_callback` is a callback function also in the `update` method.
It takes arguments `i, t, y, t_hist, ode_hist`,
where `i` is the index of `(t, y)` in `(t_hist, ode_hist)`,
and returns a dictionary that contains information to be recorded during
`update`.