Block direct call of record method of inner Logger
Inner ``Logger``s was designed to record inner data of ``BaseEnv``,
so attempts to directly record any data to the inner logger should
be blocked.

I've made a method ``Logger._record`` that works the same as the
previous ``Logger.record`` method, and add a line to check the
``Logger`` is inner or not inside the previous ``Logger.record``
method for compatibility.