Add copy and update with prune method to `parser`
The `parser.copy` method is nothing but a deepcopy with encoding.

The `prune` option in `parser.update` method is designed
to update the first argument (dict-like) with
the second arguemnt (also dict-like) removing the keys
that only exist in the first arguemnt.

This option is quite useful for resetting parsed dictionaries to their
defaults while preserving references.