Remove the registered attribute
When registering a `BaseEnv` or `BaseSystem`, instead of registering it to
`self.__dict__`, the Fym saves it to `self._systems_dict`.
This behavior is intended to distinguish the systems and normal
attiributes.

However, the previous code does not delete the existing attribute after
registering it to `self._systems_dict`.
This glitch is fixed in this commit.